==> 29 RecSys/1 Data_validation.ipynb <==
# Metrics, validation strategies and baselines

В данном jupyter notebook рассматриваются примеры того, какие схемы валидации и метрики используются в рекомендательных системах.
Также построим простые модели (бейзлайны) на данных МТС Библиотеки. 

* [Preprocessing](#preprocessing)
* [General remarks](#general-remarks)
* [Metrics](#metrics)
    * [Regression](#regression)
    * [Classification](#classification)
    * [Ranking](#ranking)
* [Validation strategies](#validation)
* [Baselines](#baselines)

import os
import numpy as np 
import pandas as pd 
from itertools import islice, cycle
from more_itertools import pairwise



<a id="preprocessing"></a>
# Preprocessing

Загрузим наши данные, теперь уже с фичами, и применим знания из [pandas-scipy-for-recsys](https://www.kaggle.com/sharthz23/pandas-scipy-for-recsys)

df = pd.read_pickle('./interactions_preprocessed.pickle')
df_users = pd.read_pickle('./users_preprocessed.pickle')
df_items = pd.read_pickle('./items_preprocessed.pickle')

df_users

df_true = pd.DataFrame({
    'user_id': ['Аня',                'Боря',               'Вася',         'Вася'],
    'item_id': ['Мастер и Маргарита', '451° по Фаренгейту', 'Зеленая миля', 'Рита Хейуорт и спасение из Шоушенка'],
})
df_true

df_recs = pd.DataFrame({
    'user_id': [
        'Аня', 'Аня', 'Аня',
        'Боря', 'Боря', 'Боря',
        'Вася', 'Вася', 'Вася',
    ],
    'item_id': [
        'Отверженные', 'Двенадцать стульев', 'Герои нашего времени',
        '451° по Фаренгейту', '1984', 'О дивный новый мир',
        'Десять негритят', 'Искра жизни', 'Зеленая миля',
    ],
    'rank': [
        1, 2, 3,
        1, 2, 3,
        1, 2, 3,
    ]
})
df_recs

df_true.set_index(['user_id', 'item_id']).join(df_recs.set_index(['user_id', 'item_id']), how='left')


df_merged = df_true.merge(df_recs, on = ['user_id','item_id'], how = 'left')

df_merged['hit@2'] = df_merged['rank'] <= 2

df_merged

df_merged['hit@2/2'] = df_merged['hit@2'] / 2
df_merged

df_prec2 = df_merged.groupby('user_id').agg({'hit@2/2':'sum'}).reset_index()


df_prec2

print('Precision@2' ,df_prec2['hit@2/2'].mean())

for k in [1, 2, 3]:
    hit_k = f'hit@{k}'
    hit_k2 = f'hit@{k}/{k}'
    df_merged[hit_k] = df_merged['rank'] <= k
    df_merged[hit_k2] = df_merged[hit_k] / k

users_count = df_merged['user_id'].nunique()

print( 'Precision@1', df_merged['hit@1/1'].sum() / users_count,'\n',
    'Precision@2', df_merged['hit@2/2'].sum() / users_count,'\n',
       'Precision@3', df_merged['hit@3/3'].sum() / users_count
       )

df_merged

df_merged['hit@2'].sum()/df_merged['hit@2'].count()

df_merged = df_merged.merge(df_merged.groupby(['user_id']).agg({'item_id':'count'}).reset_index().rename(columns = {'item_id':'users_item_count'}) , on = 'user_id', how = 'left')

df_merged

print( (df_merged['hit@1'] / df_merged['users_item_count']).sum() / users_count)

print( (df_merged['hit@2'] / df_merged['users_item_count']).sum() / users_count)

print( (df_merged['hit@3'] / df_merged['users_item_count']).sum() / users_count)

Mean Average Precision
То есть MAP - это усреднение AveragePrecision по всем пользователям. А AveragePrecision в свою очередь, это средний Precision@K по релевантным объектам одного пользователя

df_true = pd.DataFrame({
    'user_id': ['Аня',                'Боря',               'Вася',         'Вася'],
    'item_id': ['Мастер и Маргарита', '451° по Фаренгейту', 'Зеленая миля', 'Рита Хейуорт и спасение из Шоушенка'],
})
df_true

df_recs = pd.DataFrame({
    'user_id': [
        'Аня', 'Аня', 'Аня',
        'Боря', 'Боря', 'Боря',
        'Вася', 'Вася', 'Вася',
    ],
    'item_id': [
        'Отверженные', 'Двенадцать стульев', 'Герои нашего времени',
        '451° по Фаренгейту', '1984', 'О дивный новый мир',
        'Десять негритят', 'Рита Хейуорт и спасение из Шоушенка', 'Зеленая миля',
    ],
    'rank': [
        1, 2, 3,
        1, 2, 3,
        1, 2, 3,
    ]
})
df_recs

df_merged = df_true.merge(df_recs,on = ['user_id', 'item_id'], how='left')


df_merged.sort_values(by=['user_id', 'rank'] , inplace= True)

df_merged['reciprocal_rank'] = 1 / df_merged['rank']

df_merged = df_merged.reset_index()

df_merged[['reciprocal_rank']] = df_merged[['reciprocal_rank']].fillna(0)

df_merged

mrr = df_merged.groupby(['user_id']).agg({'reciprocal_rank':'max'}).reset_index()

mrr
