==> 29 RecSys/3 Baselines_by_user_age.ipynb <==
# Metrics, validation strategies and baselines

В данном jupyter notebook рассматриваются примеры того, какие схемы валидации и метрики используются в рекомендательных системах.
Также построим простые модели (бейзлайны) на данных МТС Библиотеки. 

* [Preprocessing](#preprocessing)
* [General remarks](#general-remarks)
* [Metrics](#metrics)
    * [Regression](#regression)
    * [Classification](#classification)
    * [Ranking](#ranking)
* [Validation strategies](#validation)
* [Baselines](#baselines)

import os
import numpy as np 
import pandas as pd 
from itertools import islice, cycle
from more_itertools import pairwise



import math

def is_nan(value):
    return math.isnan(float(value))

def isNaN(string):
    return string != string

<a id="preprocessing"></a>
# Preprocessing

Загрузим наши данные, теперь уже с фичами, и применим знания из [pandas-scipy-for-recsys](https://www.kaggle.com/sharthz23/pandas-scipy-for-recsys)

df = pd.read_pickle('./interactions_preprocessed.pickle')
df_users = pd.read_pickle('./users_preprocessed.pickle')
df_items = pd.read_pickle('./items_preprocessed.pickle')

test_dates = df['start_date'].unique()[-7:]



test_dates = list(pairwise(test_dates))
test_dates

split_dates = test_dates[0]

train = df[df['start_date'] < split_dates[0]]

test = df[(df['start_date'] >= split_dates[0]) & (df['start_date'] < split_dates[1])]

test = test[(test['rating'] >= 4) | (test['rating'].isnull())]

split_dates, train.shape, test.shape


<a id="baselines"></a>
# Baselines

Будем строить бейзлайн по популярному в зависимости от возраста пользователя

df_users

train.head(2)

test.head(2)

class PopularRecommender():
    def __init__(self, max_K=10, days=30, item_column='item_id', dt_column='date' , df_u = df_users):
        self.max_K = max_K
        self.days = days
        self.item_column = item_column
        self.dt_column = dt_column
        self.N = 10
        self.recommendations = []
        self.min_date = None
        self.df_u = df_u
        self.df_res = None


    def fit(self, df, ):
        min_date = df[self.dt_column].max().normalize() - pd.DateOffset(days=self.days)
        self.min_date = min_date
        self.standart_recommendations = df.loc[df[self.dt_column] > min_date, self.item_column].value_counts().head(self.max_K).index.values

    def super_fit(self, df, ):
        min_date = df[self.dt_column].max().normalize() - pd.DateOffset(days=self.days)
        self.min_date = min_date
        self.df_res = df.loc[df[self.dt_column] > min_date]\
            .merge(self.df_u , how = 'left' , on ='user_id')\
            .groupby(['age','sex','item_id']).agg({'user_id' : 'count'}).reset_index()\
            .rename(columns = {'user_id':'cnt'})\
            .sort_values(['age','sex','cnt'] , ascending=False)\
            .groupby(['age','sex']).head(self.max_K)
        self.standart_recommendations = df.loc[df[self.dt_column] > min_date, self.item_column].value_counts().head(self.max_K).index.values


    def recommend(self, users=None,):
        recs = self.standart_recommendations[:self.N]
        if users is None:
            return recs
        else:
            return list(islice(cycle([recs]), len(users)))

    def super_recommend(self, user,):
        x = self.df_u.loc[self.df_u['user_id'] == user]
        if x.empty:
            return self.standart_recommendations
        elif(is_nan(x['sex'].iloc[0])):
            return self.standart_recommendations
        elif(isNaN(x['age'].iloc[0])):
            return self.standart_recommendations
        else:
            #m = self.df_res.loc[(self.df_res['age'] == x['age'].iloc[0])]
            #m = self.df_res.loc[(self.df_res['sex'] == x['sex'].iloc[0])]
            m = self.df_res.loc[(self.df_res['age'] == x['age'].iloc[0]) & (self.df_res['sex'] == x['sex'].iloc[0])]
            n = m['item_id'].values
            return n



pop_model = PopularRecommender(days=77, dt_column='start_date')

pop_model.super_fit(train)

item_titles = dict(zip(df_items.id,df_items.title))

print(item_titles.get(271690))

top100_recs = pop_model.super_recommend(30)

top100_recs

list(map(item_titles.get, top100_recs))

# create datframe with user_id fo recs
recs = pd.DataFrame({'user_id': test['user_id'].unique()})



recs.head(2)

def super_rec( u_id, pop_model = pop_model ):
    g = pop_model.super_recommend(u_id)
    return g

super_rec(34)

recs['item_id'] = recs['user_id'].apply(super_rec)

recs

recs = recs.explode('item_id')
recs.head(top_N + 2)

recs['rank'] = recs.groupby('user_id').cumcount() + 1
recs.head(top_N + 2)

recs.rename(columns={'rank':'predict_rank'} , inplace = True)

recs.head(2)

test.head(2)


test_recs = test.merge(recs, how = 'left' , on = ['user_id','item_id'])

test_recs.head(2)

test_recs = test.merge(recs, how = 'left' , on = ['user_id','item_id'])

uic = test_recs.groupby('user_id').agg({'item_id':'count'}).reset_index().rename(columns = {'item_id':'users_item_count'})

test_recs = test_recs.merge(uic , how = 'left' , on = 'user_id')

test_recs['pred_reciprocal_rank'] = 1 / test_recs['predict_rank']
test_recs['pred_reciprocal_rank'] = test_recs['pred_reciprocal_rank'].fillna(0)

test_recs.head(2)

test_recs = test_recs.sort_values(by=['user_id', 'predict_rank'])



test_recs['cumulative_rank'] = test_recs.groupby('user_id').cumcount() + 1

test_recs

test_recs['cumulative_rank'] = test_recs['cumulative_rank'] / test_recs['predict_rank']

test_recs.head(2)

test_recs[test_recs['predict_rank'].notnull()]

print(f'Метрик по test ({str(split_dates[0])[:10]}, {str(split_dates[1])[:10]})')

users_count = test_recs['user_id'].nunique()

users_count

for k in range(1, top_N + 1):
    hit_k = f'hit@{k}'
    test_recs[hit_k] = test_recs['predict_rank'] <= k
    print(f'Precision@{k} = {(test_recs[hit_k] / k).sum() / users_count:.4f}')
    print(f"Recall@{k} = {(test_recs[hit_k] / test_recs['users_item_count']).sum() / users_count:.4f}")

mapN = (test_recs["cumulative_rank"] / test_recs["users_item_count"]).sum() / users_count
print(f"MAP@{top_N} = {mapN}")

mrr = test_recs.groupby('user_id').agg({'pred_reciprocal_rank':'max'}).reset_index()['pred_reciprocal_rank'].mean()
print(f"MRR = {mrr}")
