==> 29 RecSys/4 Embedding-Based Recommender.ipynb <==
import tensorflow_datasets as tfds
import tensorflow as tf
import pandas as pd

df = pd.read_parquet('df_embed.parquet.gzip')

df

filtered_data = (
    df.filter(["timestamp", "user_id", "movie_id", "user_rating"])
        .sort_values("timestamp")
        .astype({"user_id": int, "movie_id": int, "user_rating": int})
        .drop(columns=["timestamp"])
)

train = filtered_data.iloc[:900000]
test = filtered_data.iloc[900000:]

train

print(train.query("user_id == 1").shape[0])

print(test.query("user_id == 1").shape[0])

test.query("user_id == 1")

X_train = train.drop(columns=["user_rating"])
y_train = train["user_rating"]
X_test = test.drop(columns=["user_rating"])
y_test = test["user_rating"]

##BAD BAD BAD
from sklearn.ensemble import HistGradientBoostingRegressor
from sklearn.metrics import mean_absolute_error, r2_score

hgb = HistGradientBoostingRegressor(random_state=0)

type(hgb)

hgb.fit(X_train, y_train)

print(hgb.score(X_test, y_test))

print(hgb.score(X_train, y_train))

y_pred = hgb.predict(X_test)

y_pred

print(mean_absolute_error(y_test, y_pred))

Implementation in TensorFlow, Version One

all_users = train["user_id"].unique()
all_movies = train["movie_id"].unique()

all_users

all_movies

import tensorflow as tf

user_input = tf.keras.layers.Input(shape=(1,), name="user")
user_as_integer = tf.keras.layers.IntegerLookup(vocabulary=all_users)(user_input)
user_embedding = tf.keras.layers.Embedding(input_dim=len(all_users) + 1, output_dim=32)(user_as_integer)

movie_input = tf.keras.layers.Input(shape=(1,), name="movie")
movie_as_integer = tf.keras.layers.IntegerLookup(vocabulary=all_movies)(movie_input)
movie_embedding = tf.keras.layers.Embedding(input_dim=len(all_movies) + 1, output_dim=32)(movie_as_integer)

dot = tf.keras.layers.Dot(axes=2)([user_embedding, movie_embedding])
flatten = tf.keras.layers.Flatten()(dot)

model = tf.keras.Model(inputs=[user_input, movie_input], outputs=flatten)

model.compile(loss="mse", metrics=[tf.keras.metrics.MeanAbsoluteError()])

model.fit(
    x={
        "user": X_train["user_id"],
        "movie": X_train["movie_id"]
    },
    y=y_train.values,
    batch_size=256,
    epochs=100,
    validation_split=0.1, # for early stopping
    callbacks=[
        tf.keras.callbacks.EarlyStopping(patience=1, restore_best_weights=True)
    ],
)


type(user_input)
